+++ tests/framework/db/pgsql/schema/SchemaTest.php
@@ -21,7 +21,7 @@ final class SchemaTest extends \yiiunit\framework\db\schema\AbstractSchema
 
         $this->columnsSchema = match (version_compare($this->db->serverVersion, '10.0', '>=')) {
             true => [
-                'id' => 'INT GENERATED ALWAYS AS IDENTITY',
+                'id' => 'INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY',
                 'name' => 'VARCHAR(128)',
             ],
             default => [
@@ -29,5 +29,17 @@ final class SchemaTest extends \yiiunit\framework\db\schema\AbstractSchema
                 'name' => 'VARCHAR(128)',
             ],
         };
+    }
+
+    /**
+     * @dataProvider \yiiunit\framework\db\pgsql\provider\SchemaProvider::resetSequence
+     */
+    public function testResetSequence(
+        string $tableName,
+        array $insertRows,
+        array $expectedIds,
+        int|null $value = null
+    ): void {
+        parent::testResetSequence($tableName, $insertRows, $expectedIds, $value);
     }
 }