+++ tests/framework/db/schema/AbstractTableSchema.php
@@ -33,10 +33,10 @@ abstract class AbstractTableSchema extends \yiiunit\TestCase
         };
 
         // MySQL, SQLite does not support sequences
-        match ($this->db->getDriverName()) {
+        match ($this->db->driverName) {
             'oci' => $this->assertStringContainsString('ISEQ$$_', $tableSchema->sequenceName),
             'pgsql' => $this->assertSame('T_sequence_name_id_seq', $tableSchema->sequenceName['id']),
-            default => $this->assertEmpty($tableSchema->sequenceName),
+            default => $this->assertSame('id', $tableSchema->sequenceName),
         };
 
         $this->ensureNoTable($tableName);
@@ -71,7 +71,7 @@ abstract class AbstractTableSchema extends \yiiunit\TestCase
         match ($this->db->getDriverName()) {
             'oci' => $this->assertStringContainsString('ISEQ$$_', $tableSchema->sequenceName),
             'pgsql' => $this->assertSame('T_sequence_name_with_pk_id_seq', $tableSchema->sequenceName['id']),
-            default => $this->assertEmpty($tableSchema->sequenceName),
+            default => $this->assertSame('id', $tableSchema->sequenceName),
         };
 
         $this->ensureNoTable($tableName);
@@ -95,7 +95,8 @@ abstract class AbstractTableSchema extends \yiiunit\TestCase
         // MySQL, SQLite does not support sequences
         match ($this->db->driverName) {
             'oci' => $this->assertStringContainsString('ISEQ$$_', $tableSchema->sequenceName),
-            default => $this->assertEmpty($tableSchema->sequenceName),
+            'sqlite' => $this->assertEmpty($tableSchema->sequenceName),
+            default => $this->assertSame('id1', $tableSchema->sequenceName),
         };
 
         $this->ensureNoTable($tableName);