+++ tests/framework/db/sqlite/provider/SchemaProvider.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace yiiunit\framework\db\sqlite\provider;
+
+final class SchemaProvider extends \yiiunit\framework\db\provider\AbstractSchemaProvider
+{
+}